feat(fixed_term_deposit): update fixed term deposit screen

+++ lib/features/fixed_term_deposit/presentation/screens/fixed_term_deposit_screen.dart
@@ -1,4 +1,6 @@
+import 'package:expense_control_app/core/presentation/widgets/action_button_widget.dart';
 import 'package:expense_control_app/features/fixed_term_deposit/presentation/state/fixed_term_deposit_notifier.dart';
+import 'package:expense_control_app/features/fixed_term_deposit/presentation/widgets/fixed_term_deposit_alert_dialog_form/fixed_term_deposit_alert_dialog_form_add_widget.dart';
 import 'package:expense_control_app/features/fixed_term_deposit/presentation/widgets/fixed_term_deposit_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -31,18 +33,30 @@ class FixedTermDepositScreen extends ConsumerWidget {
                       children: [
                         SlidableAction(
                           onPressed: (context) async {
-                            await ref
-                                .read(fixedTermDepositProvider.notifier)
-                                .updateFixedTermDeposit(
-                                  id: state.fixedTermDeposits[index].id,
-                                  depositAmount: 6,
-                                  depositAmountReceived: 100,
-                                  depositDate: DateTime(2000, 1, 1),
-                                  depositDueDate: DateTime(2000, 1, 6),
-                                  dolarPrice: 1.0,
-                                  name: 'Depósito actualizado',
-                                );
-                            ref.invalidate(fixedTermDepositProvider);
+                            // await showDialog(
+                            //   context: context,
+                            //   builder: (context) {
+                            //     return FixedTermDepositAlertDialogFormUpdateWidget(
+                            //       id: state.fixedTermDeposits[index].id,
+                            //       depositAmount: state
+                            //           .fixedTermDeposits[index]
+                            //           .depositAmount,
+                            //       depositAmountReceived: state
+                            //           .fixedTermDeposits[index]
+                            //           .depositAmountReceived,
+                            //       depositDate: state
+                            //           .fixedTermDeposits[index]
+                            //           .depositDate,
+                            //       depositDueDate: state
+                            //           .fixedTermDeposits[index]
+                            //           .depositDueDate,
+                            //       dolarPrice:
+                            //           state.fixedTermDeposits[index].dolarPrice,
+                            //       fixedTermDepositName:
+                            //           state.fixedTermDeposits[index].name,
+                            //     );
+                            //   },
+                            // );
                           },
                           backgroundColor: Colors.blue,
                           foregroundColor: Colors.white,
@@ -57,12 +71,46 @@ class FixedTermDepositScreen extends ConsumerWidget {
                       children: [
                         SlidableAction(
                           onPressed: (context) async {
-                            await ref
-                                .read(fixedTermDepositProvider.notifier)
-                                .deleteFixedTermDeposit(
-                                  id: state.fixedTermDeposits[index].id,
+                            await showDialog(
+                              context: context,
+                              builder: (context) {
+                                return AlertDialog(
+                                  title: const Center(
+                                    child: Text(
+                                      '¿Segura que quiere eliminar este gasto?',
+                                    ),
+                                  ),
+                                  actions: [
+                                    ActionButtonWidget(
+                                      onPressed: () async {
+                                        Navigator.pop(context);
+                                      },
+                                      text: 'Cancelar',
+                                    ),
+                                    ActionButtonWidget(
+                                      onPressed: () async {
+                                        await ref
+                                            .read(
+                                              fixedTermDepositProvider.notifier,
+                                            )
+                                            .deleteFixedTermDeposit(
+                                              id: state
+                                                  .fixedTermDeposits[index]
+                                                  .id,
+                                            );
+                                        ref.invalidate(
+                                          fixedTermDepositProvider,
+                                        );
+                                        if (context.mounted) {
+                                          Navigator.pop(context);
+                                        }
+                                      },
+                                      text: 'Confirmar',
+                                    ),
+                                  ],
                                 );
-                            ref.invalidate(fixedTermDepositProvider);
+                              },
+                            );
                           },
                           backgroundColor: Colors.red,
                           foregroundColor: Colors.white,
@@ -82,28 +130,6 @@ class FixedTermDepositScreen extends ConsumerWidget {
                       depositDueDate:
                           state.fixedTermDeposits[index].depositDueDate,
                       dolarPrice: state.fixedTermDeposits[index].dolarPrice,
-                      onDeleted: () async {
-                        await ref
-                            .read(fixedTermDepositProvider.notifier)
-                            .deleteFixedTermDeposit(
-                              id: state.fixedTermDeposits[index].id,
-                            );
-                        ref.invalidate(fixedTermDepositProvider);
-                      },
-                      onEdited: () async {
-                        await ref
-                            .read(fixedTermDepositProvider.notifier)
-                            .updateFixedTermDeposit(
-                              id: state.fixedTermDeposits[index].id,
-                              depositAmount: 6,
-                              depositAmountReceived: 100,
-                              depositDate: DateTime(2000, 1, 1),
-                              depositDueDate: DateTime(2000, 1, 6),
-                              dolarPrice: 1.0,
-                              name: 'Depósito actualizado',
-                            );
-                        ref.invalidate(fixedTermDepositProvider);
-                      },
                     ),
                   );
                 },
@@ -114,17 +140,12 @@ class FixedTermDepositScreen extends ConsumerWidget {
       ),
       floatingActionButton: FloatingActionButton.extended(
         onPressed: () async {
-          await ref
-              .read(fixedTermDepositProvider.notifier)
-              .addFixedTermDeposit(
-                depositAmount: 6,
-                depositAmountReceived: 100,
-                depositDate: DateTime(2000, 1, 1),
-                depositDueDate: DateTime(2000, 1, 30),
-                dolarPrice: 1.0,
-                name: 'Depósito actualizado',
-              );
-          ref.invalidate(fixedTermDepositProvider);
+          await showDialog(
+            context: context,
+            builder: (context) {
+              return const FixedTermDepositAlertDialogFormAddWidget();
+            },
+          );
         },
         label: const Row(
           children: [
